# university-mgmt-service/src/main/java/com/university/management/controller/CourseController.java
package com.university.management.controller;

import com.university.management.common.api.CoursesApi;
import com.university.management.common.model.CourseCreateRequest;
import com.university.management.common.model.CourseResponse;
import com.university.management.service.service.CourseManagementService;
import java.util.List;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1")
public class CourseController implements CoursesApi {

  private final CourseManagementService courseManagementService;

  public CourseController(CourseManagementService courseManagementService) {
    this.courseManagementService = courseManagementService;
  }

  @Override
  public ResponseEntity<CourseResponse> createCourse(CourseCreateRequest courseCreateRequest) {
    return ResponseEntity.status(HttpStatus.CREATED)
        .body(courseManagementService.createCourse(courseCreateRequest));
  }

  @Override
  public ResponseEntity<CourseResponse> getCourse(UUID courseId) {
    return ResponseEntity.ok(courseManagementService.getCourse(courseId));
  }

  @Override
  public ResponseEntity<List<CourseResponse>> listCourses() {
    return ResponseEntity.ok(courseManagementService.listCourses());
  }
}